+++ account-api/src/main/java/com/ecommerce/accountservice/account/api/AccountController.java
@@ -1,0 +1,39 @@
+package com.ecommerce.accountservice.account.api;
+
+import com.ecommerce.accountservice.account.service.AccountService;
+import com.ecommerce.accountservice.account.dto.AccountResponseDto;
+import com.ecommerce.accountservice.account.dto.SignUpRequestDto;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/accounts")
+@Slf4j
+public class AccountController {
+
+    private final AccountService accountService;
+
+    @PostMapping("/sign-up")
+    public ResponseEntity<?> signup(@RequestBody SignUpRequestDto request) {
+        AccountResponseDto response = accountService.createAccount(request);
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
+    }
+
+    @GetMapping("/{accountId}")
+    public ResponseEntity<?> findSimpleAccountInfo(@PathVariable Long accountId, HttpServletRequest httpServletRequest) {
+        Long requesterId = Long.valueOf(httpServletRequest.getHeader("X-User-Id"));
+        AccountResponseDto response = accountService.findSimpleAccountInfoById(accountId);
+        return ResponseEntity.status(HttpStatus.OK).body(response);
+    }
+
+    @GetMapping("/health-check")
+    public ResponseEntity<?> healthCheck() {
+        log.info("health check...");
+        return ResponseEntity.ok("health check...");
+    }
+}
