+++ account-api/src/main/java/com/ecommerce/accountservice/account/dto/SignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.ecommerce.accountservice.api.dto;
+package com.ecommerce.accountservice.account.dto;
 
 import lombok.Builder;
 import lombok.Getter;
