+++ account-api/src/main/java/com/ecommerce/accountservice/account/repository/AccountRepository.java
@@ -1,6 +1,6 @@
-package com.ecommerce.accountservice.domain.repo;
+package com.ecommerce.accountservice.account.repository;
 
-import com.ecommerce.accountservice.domain.Account;
+import com.ecommerce.accountservice.account.entity.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountRepository extends JpaRepository<Account, Long> {
