+++ account-api/src/main/java/com/ecommerce/accountservice/kafka/service/consumer/KafkaConsumerService.java
@@ -1,6 +1,6 @@
 package com.ecommerce.accountservice.kafka.service.consumer;
 
-import com.ecommerce.accountservice.domain.service.CouponService;
+import com.ecommerce.accountservice.account.service.CouponService;
 import com.ecommerce.accountservice.kafka.config.TopicConfig;
 import com.ecommerce.accountservice.kafka.dto.CouponIssuanceResultKafkaEvent;
 import lombok.RequiredArgsConstructor;
@@ -8,8 +8,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
-
-import java.util.Objects;
 
 @Service @Slf4j
 @RequiredArgsConstructor
