+++ account-api/src/test/java/com/ecommerce/accountservice/AccountServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.nayoung.accountservice;
+package com.ecommerce.accountservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
